requirements.txt
# ==================== LangGraph 核心依赖 ====================
langgraph>=0.2.70
langchain>=1.0.5
langchain-core>=1.0.4

# ==================== LLM Provider ====================
langchain-openai>=0.3.0

# ==================== SQLite 持久化 ====================
langgraph-checkpoint-sqlite>=1.0.0

# ==================== 环境变量管理 ====================
python-dotenv>=1.0.0

# ==================== Flask Web 服务 ====================
flask>=3.0.0
flask-cors>=4.0.0
